+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/Server/Miscelleanous.lua
@@ -100,3 +100,11 @@ Ext.RegisterOsirisListener("ObjectTurnStarted", 1, "before", function(object)
     end
 end)
 ----------
+---------- Last Rites workaround
+Ext.RegisterOsirisListener("CharacterUsedSkill", 4, "before", function(character, skill, skillType, skillElement)
+    if skill == "Teleportation_LastRites" then
+        ---@type EsvCharacter
+        local char = Ext.GetCharacter(character)
+        char.Stats.CurrentVitality = 1
+    end
+end)