+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/Shared/StatsPatching.lua
@@ -230,9 +230,21 @@ local function TeleportRevert()
 	local hardSaved = Ext.LoadFile("LeaderLib_GlobalSettings.json")
 	if hardSaved ~= nil then
 		local variables = Ext.JsonParse(hardSaved).Mods["3ff156e2-289e-4dac-81f5-a44e3e304163"].Global.Flags
-		if variables.LXDGM_ModuleOriginalTeleport.Enabled then
+		if variables.LXDGM_ModuleOriginalTeleport and variables.LXDGM_ModuleOriginalTeleport.Enabled then
 			Ext.Print("Reverting Teleport and Nether Swap targetting conditions...")
 			Ext.AddPathOverride("Public/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Stats/Generated/Data/LX_TeleportNS.txt", "Public/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Stats/Generated/Data/LX_Empty.txt")
+		end
+	end
+end
+
+local function FixPhysicalResistance()
+	if Ext.Version() <= 55 then
+		for i,stat in pairs(Ext.GetStatEntries("Character")) do
+			stat = Ext.GetStat(stat)
+			local physicalResistance = stat.PhysicalResistance
+			if physicalResistance > 0 then
+				Ext.StatSetAttribute(stat.Name, "PhysicalResistance", math.floor(physicalResistance/2))
+			end
 		end
 	end
 end
@@ -247,4 +259,5 @@ Ext.RegisterListener("StatsLoaded", ReplaceDescriptionParams)
 Ext.RegisterListener("StatsLoaded", AdjustNPCStats)
 Ext.RegisterListener("StatsLoaded", CustomScalings)
 Ext.RegisterListener("StatsLoaded", ChameleonCloakRevert)
+Ext.RegisterListener("StatsLoaded", FixPhysicalResistance)
 Ext.RegisterListener("ModuleLoadStarted", TeleportRevert)